backend/requirements.txt
# ─── Web framework ───
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
slowapi>=0.1.9             # Rate limiting (brug: @limiter.limit("10/minute"))

# ─── Database ───
sqlalchemy==2.0.35
alembic>=1.13.0        # Database-migrationer med versionshistorik og rollback
pyodbc>=5.0.0          # SQL Server i produktion (kræver ODBC Driver 18 for SQL Server)

# ─── Auth ───
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# ─── Validering ───
pydantic[email]==2.9.0  # [email] kræves til EmailStr

# ─── Konfiguration ───
python-dotenv==1.0.1

# ─── AI / Ollama ───
requests==2.32.0

# ─── Transskribering (CoRal Røst v2) ───
transformers==4.46.0
torch==2.5.0
numpy>=1.26.0
av>=12.0.0              # PyAV: bundlet ffmpeg-bindings — håndterer webm/opus/mp3 uden ekstern ffmpeg

# ─── Test ───
pytest>=8.0
httpx>=0.27             # AsyncClient til FastAPI TestClient
pytest-cov>=5.0         # Coverage-rapportering